requirements.txt
requests
python-telegram-bot==20.3
cryptography
pytest
